requirements.txt
aiogram==3.19.0
openai>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
aiofiles>=23.0.0 